# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Newsletter API",
    "description": "Newsletter's API for authentication, users and news management",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "paths": {
    "/api/news": {
      "get": {
        "summary": "Returns list of news",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "category": { "type": "string" },
                "text": { "type": "string" },
                "date": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/api/news/add": {
      "post": {
        "security": [{ "BasicAuth": ["admin"] }],
        "summary": "Creates a news piece",
        "parameters": [
          {
            "in": "body",
            "name": "category",
            "required": true,
            "type": "string",
            "description": "Title/category of the news piece"
          },
          {
            "in": "body",
            "name": "text",
            "required": true,
            "type": "string",
            "description": "Content of the news piece"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad request"
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" }
              }
            }
          }
        }
      }
    },
    "/api/news/del": {
      "post": {
        "security": [{ "BasicAuth": ["admin"] }],
        "summary": "Deletes a news piece",
        "parameters": [
          {
            "in": "body",
            "name": "ids",
            "required": true,
            "type": "array",
            "description": "List of news pieces' ids to delete"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/pass": {
      "post": {
        "security": [{ "BasicAuth": ["user", "admin"] }],
        "summary": "Changes password",
        "parameters": [
          {
            "in": "body",
            "name": "oldPassword",
            "required": true,
            "type": "string",
            "description": "User's old password"
          },
          {
            "in": "body",
            "name": "newPassword",
            "required": true,
            "type": "string",
            "description": "User's new password"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth": {
      "post": {
        "summary": "Authenticates user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "username",
            "required": true,
            "type": "string",
            "description": "Username"
          },
          {
            "in": "body",
            "name": "password",
            "required": true,
            "type": "string",
            "description": "Password"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "securityDefintions": {
    "BasicAuth": {
      "type": "basic",
      "in": "header",
      "name": "x-auth-token",
      "scopes": {
        "admin": "Can create, read, update and delete both news and users",
        "user": "Can create, read, update and delete news"
      }
    }
  }
}